move Fluxx client to separate file
