+++ exportScript.py
@@ -8,12 +8,10 @@ from datetime import datetime
 import boto3
 import django
 import paramiko
-import requests
 from botocore.exceptions import NoCredentialsError
 
-from exporter.models import Column, Entity
-from fetchConfigDetails import (fetch_fluxx_configs, fetch_s3_configs,
-                                fetch_sftp_configs)
+from exporter.clients import FluxxClient
+from fetchConfigDetails import fetch_s3_configs, fetch_sftp_configs
 
 django.setup()
 
@@ -40,7 +38,7 @@ def main():
         print(f"inputs: {inputs}")
         inputs = parse_data(inputs)
         print(f"inputs: {inputs}")
-        fluxx = Fluxx()
+        fluxx = FluxxClient()
 
         for entity_data in inputs:
             entity = str(entity_data[0]).lower()
@@ -336,131 +334,6 @@ def upload_directory_to_s3(directory, bucket, access_key, secret_key):
                 print("Credentials not available")
             except Exception as e:
                 print(f"Failed to upload {file_name} to {bucket}: {e}")
-
-
-class Fluxx(object):
-    '''Fluxx object that will handle connection/session and subsequent requests'''
-
-    version = 'v2'
-
-    def __init__(self, config_name='Armanino Demo'):
-        # get config information from DB
-        instance_domain, application_id, secret = fetch_fluxx_configs(
-            config_name)
-
-        # Optional style parameter, generally best kept as 'Full'
-        # style = 'full'
-
-        # concat domain and extension onto API request/not accounting for
-        # Pre-production environments (can be added)
-        domain = 'fluxx'
-        extension = 'io'
-
-        # create base URL for fluxx
-        fluxx_url = f'https://{instance_domain}.{domain}.{extension}/'
-
-        # generate token
-        token_url = fluxx_url + 'oauth/token'
-
-        # oauth parameters to retrieve token
-        oauth_params = {
-            'grant_type': 'client_credentials',
-            'client_id': application_id,
-            'client_secret': secret
-        }
-
-        # begin a Request Session
-        self.session = requests.Session()
-
-        # obtain oauth token
-        try:
-            response = self.session.post(token_url, data=oauth_params)
-        except BaseException:
-            print("No response from client API")
-
-        # Set Session request headers to persist connection
-        try:
-            self.token = response.json()['access_token']
-            self.session.headers.update({
-                'Authorization': f'Bearer {self.token}'
-            })
-        except BaseException:
-            print("Could not find access token")
-
-        self.api_url = fluxx_url + 'api/rest/v2/'
-
-    def list_rows(self, entity, columns, page=1, per_page=100,
-                  filter=None, related_entity=None):
-        '''Function to return a list of some number of rows and pages relating to an entity (or table)
-        Filters can be applied following the format: <entity> <logic> <condition>
-        Example: amount_requested eq 10000
-        Filters can be tied together as follows: amount_requested eq 10000 and created_at today
-        Filter logic will be requested and further documented at a later date
-        '''
-
-        if page < 1:
-            raise ValueError("Page integer must be greater than 0.")
-        print(f"columns: {json.dumps(columns)}")
-        list_params = {
-            'cols': json.dumps(columns),
-            'page': page,
-            'per_page': per_page
-        }
-
-        if filter:
-            list_params.update({
-                'filter': json.dumps(filter)
-            })
-
-        if related_entity:
-            entity_id = Entity.objects.get(name=entity).id
-
-            related_entity_objects = Entity.objects.filter(
-                related_entity=entity_id)
-
-            for related_entity in related_entity_objects:
-                related_entity_id = Entity.objects.get(name=related_entity).id
-                print(f"related_entity: {related_entity.name}")
-                re_columns = [
-                    column.name for column in Column.objects.filter(
-                        entity=related_entity_id)]
-                print(f"re_columns: {re_columns}")
-
-                # Directly assign the list to relation_params
-                relation_params = {related_entity.name: re_columns}
-
-                list_params.update({
-                    # Directly assign the list to relation_params
-                    'relation': json.dumps(relation_params),
-                })
-
-        print(f"{list_params}")
-
-        return self.session.get(self.api_url + entity, params=list_params)
-
-    def download_document(self, document_id, sub_folder_path):
-        document_id = str(document_id)
-        download_params = {
-            'cols': json.dumps(["document_file_name"])
-        }
-        # Get the filename and type
-        document_info = self.session.get(
-            self.api_url + 'model_document/' + document_id,
-            params=download_params)
-        document_name = document_info.json(
-        )['model_document']['document_file_name']
-
-        # Get the file and save to respective folder
-        document_download_url = self.api_url + 'model_document_download/' + document_id
-        response = self.session.get(document_download_url, stream=True)
-
-        # Ensure the Documents directory exists
-        documents_path = os.path.join(sub_folder_path, 'Documents')
-        os.makedirs(documents_path, exist_ok=True)
-
-        with open(os.path.join(documents_path, document_name), 'wb') as file:
-            for chunk in response.iter_content(chunk_size=8192):
-                file.write(chunk)
 
 
 if __name__ == "__main__":
